fix: CLI Warning fixed and results.py added
